refactor(hir): take advantage of `pub use`
